requirements.txt
pandas==2.1.4
numpy==1.26.2
networkx==3.2.1
matplotlib==3.8.2
python-louvain==0.16
pyyaml==6.0.1
